Allow kind-sysbox compilation when operating in docker's userns-remap mode
Sysbox runtime can either operate in 'uid-shifting' mode (if 'shiftfs' kernel module is present), or in 'non-uid-shifting' mode. For the later to work we make use of docker's userns-remap feature.

In this operational mode, we must ensure that kind-sysbox's building process can dump its artifacts in the expected 'out-dir' directory (i.e. root-repo/bin). This can only work if the building process within the container has 'write' access to out-dir. For this to work we add simple 'chmod / chown' instructions to temporarily relax write access to this folder.

Signed-off-by: Rodny Molina <rmolina@nestybox.com>